1004335: abort with error if org names have '<>/' in them
Spacewalk is less strict about characters in role names than Katello, so we
need to check for such chars so we don't find out in the middle of org
creation.

We could have changed the '<>/' chars to '_', but it would be difficult to
change them back in the role label once Katello allows these chars in role
names.

Also, remove global warning during runs of spacewalk-splice-checkin.