use sync_users config setting
In older versions of sst, users were synced from spacewalk to katello.

This doesn't work correctly with multiple satellites however, since there can
be collisions. Instead, default to not syncing users, and only sync if
"sync_users" is set to true, AND there is only one spacewalk that is being
synced.